COMPASS-890: Undo special-casing `allowDiskUse` for Atlas (#90)
Atlas now allows the `allowDiskUse` aggregation option, silently ignoring it when it is not available. This commit undoes the changes introduced to work around `allowDiskUse` throwing an error in the
run-up to the free-tier launch.  `allowDiskUse` defaults to `true` for the aggregation call, just as we did before.

This will also fix COMPASS-920